moved ContainerNSToPID; adjusted contianerd errorhandling
