+++ utils/containers.go
@@ -1,7 +1,10 @@
 package utils
 
 import (
+	"fmt"
 	"os"
+	"path/filepath"
+	"strconv"
 	"strings"
 )
 
@@ -49,4 +52,22 @@ func GetCNIBinaryPath() string {
 		cniPath = cniBin
 	}
 	return cniPath
+}
+
+// ContainerNSToPID resolves the name of a container via
+// the "/run/netns/<CONTAINERNAME>" to its PID
+func ContainerNSToPID(cID string) (int, error) {
+	pnns, err := filepath.EvalSymlinks("/run/netns/" + cID)
+	if err != nil {
+		return 0, err
+	}
+	pathElem := strings.Split(pnns, "/")
+	if len(pathElem) != 4 {
+		return 0, fmt.Errorf("unexpected result looking up container PID")
+	}
+	pid, err := strconv.Atoi(pathElem[1])
+	if err != nil {
+		return 0, fmt.Errorf("error converting the string part of the namespace link to int")
+	}
+	return pid, nil
 }
